Implemented (extremely) basic split view controller
