--- Timeline/View/ImportPhotosViewController.xib
@@ -1,77 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.Cocoa.XIB" version="3.0" toolsVersion="13122.17" systemVersion="16G29" targetRuntime="MacOSX.Cocoa" propertyAccessControl="none" useAutolayout="YES" customObjectInstantitationMethod="direct">
-    <dependencies>
-        <deployment identifier="macosx"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.CocoaPlugin" version="13122.17"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <customObject id="-2" userLabel="File's Owner" customClass="ImportPhotosViewController" customModule="Timeline" customModuleProvider="target">
-            <connections>
-                <outlet property="view" destination="Hz6-mo-xeY" id="0bl-1N-x8E"/>
-            </connections>
-        </customObject>
-        <customObject id="-1" userLabel="First Responder" customClass="FirstResponder"/>
-        <customObject id="-3" userLabel="Application" customClass="NSObject"/>
-        <customView id="Hz6-mo-xeY">
-            <rect key="frame" x="0.0" y="0.0" width="500" height="150"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMinY="YES"/>
-            <subviews>
-                <progressIndicator wantsLayer="YES" maxValue="100" style="bar" translatesAutoresizingMaskIntoConstraints="NO" id="QzG-gC-Wgx">
-                    <rect key="frame" x="20" y="19" width="460" height="20"/>
-                </progressIndicator>
-                <textField horizontalHuggingPriority="251" verticalHuggingPriority="750" translatesAutoresizingMaskIntoConstraints="NO" id="xrN-w9-IhD">
-                    <rect key="frame" x="239" y="38" width="23" height="17"/>
-                    <textFieldCell key="cell" scrollable="YES" lineBreakMode="clipping" sendsActionOnEndEditing="YES" title="0%" id="vnr-g4-HP4">
-                        <font key="font" metaFont="system"/>
-                        <color key="textColor" name="labelColor" catalog="System" colorSpace="catalog"/>
-                        <color key="backgroundColor" name="controlColor" catalog="System" colorSpace="catalog"/>
-                    </textFieldCell>
-                </textField>
-                <textField horizontalHuggingPriority="251" verticalHuggingPriority="750" translatesAutoresizingMaskIntoConstraints="NO" id="pHf-bk-Eyg">
-                    <rect key="frame" x="238" y="55" width="25" height="17"/>
-                    <textFieldCell key="cell" scrollable="YES" lineBreakMode="clipping" sendsActionOnEndEditing="YES" title="0/0" id="C6h-8Z-Ku1">
-                        <font key="font" metaFont="system"/>
-                        <color key="textColor" name="labelColor" catalog="System" colorSpace="catalog"/>
-                        <color key="backgroundColor" name="controlColor" catalog="System" colorSpace="catalog"/>
-                    </textFieldCell>
-                </textField>
-                <textField horizontalHuggingPriority="251" verticalHuggingPriority="750" translatesAutoresizingMaskIntoConstraints="NO" id="5jg-9d-nkV">
-                    <rect key="frame" x="18" y="110" width="464" height="20"/>
-                    <constraints>
-                        <constraint firstAttribute="height" constant="20" id="xck-gl-8Pf"/>
-                    </constraints>
-                    <textFieldCell key="cell" scrollable="YES" lineBreakMode="clipping" sendsActionOnEndEditing="YES" title="Importing..." id="rzp-aG-gh9">
-                        <font key="font" metaFont="system"/>
-                        <color key="textColor" name="labelColor" catalog="System" colorSpace="catalog"/>
-                        <color key="backgroundColor" name="controlColor" catalog="System" colorSpace="catalog"/>
-                    </textFieldCell>
-                </textField>
-                <pathControl verticalHuggingPriority="750" allowsExpansionToolTips="YES" translatesAutoresizingMaskIntoConstraints="NO" id="LyQ-MI-siC">
-                    <rect key="frame" x="20" y="72" width="460" height="22"/>
-                    <pathCell key="cell" selectable="YES" editable="YES" alignment="left" id="aDd-Ng-bvx">
-                        <font key="font" metaFont="system"/>
-                        <url key="url" string="file:///Users/"/>
-                        <color key="backgroundColor" red="0.8980392157" green="0.92549019610000005" blue="0.97254901959999995" alpha="0.0" colorSpace="deviceRGB"/>
-                    </pathCell>
-                </pathControl>
-            </subviews>
-            <constraints>
-                <constraint firstItem="QzG-gC-Wgx" firstAttribute="top" secondItem="xrN-w9-IhD" secondAttribute="bottom" id="4dN-xi-jZ1"/>
-                <constraint firstAttribute="bottom" secondItem="QzG-gC-Wgx" secondAttribute="bottom" constant="20" id="6eg-dk-JLN"/>
-                <constraint firstAttribute="trailing" secondItem="5jg-9d-nkV" secondAttribute="trailing" constant="20" id="8HB-ZJ-ojc"/>
-                <constraint firstItem="5jg-9d-nkV" firstAttribute="leading" secondItem="Hz6-mo-xeY" secondAttribute="leading" constant="20" id="FcI-lE-dwF"/>
-                <constraint firstItem="LyQ-MI-siC" firstAttribute="trailing" secondItem="5jg-9d-nkV" secondAttribute="trailing" id="I0e-PB-4Dc"/>
-                <constraint firstItem="pHf-bk-Eyg" firstAttribute="top" secondItem="LyQ-MI-siC" secondAttribute="bottom" id="ITO-Lg-MO5"/>
-                <constraint firstItem="QzG-gC-Wgx" firstAttribute="leading" secondItem="Hz6-mo-xeY" secondAttribute="leading" constant="20" id="NZg-6J-CQ1"/>
-                <constraint firstItem="xrN-w9-IhD" firstAttribute="centerX" secondItem="QzG-gC-Wgx" secondAttribute="centerX" id="Rsu-Ah-id7"/>
-                <constraint firstAttribute="trailing" secondItem="QzG-gC-Wgx" secondAttribute="trailing" constant="20" id="VbL-dk-I9N"/>
-                <constraint firstItem="xrN-w9-IhD" firstAttribute="top" secondItem="pHf-bk-Eyg" secondAttribute="bottom" id="W8J-VF-m5X"/>
-                <constraint firstItem="LyQ-MI-siC" firstAttribute="leading" secondItem="5jg-9d-nkV" secondAttribute="leading" id="sPK-mu-sdw"/>
-                <constraint firstItem="5jg-9d-nkV" firstAttribute="top" secondItem="Hz6-mo-xeY" secondAttribute="top" constant="20" id="tMs-oz-T7k"/>
-                <constraint firstItem="pHf-bk-Eyg" firstAttribute="centerX" secondItem="xrN-w9-IhD" secondAttribute="centerX" id="y9F-Tz-6yo"/>
-            </constraints>
-            <point key="canvasLocation" x="116" y="106"/>
-        </customView>
-    </objects>
-</document>
