+++ Timeline/View/MainWindow/MainSplitViewController.swift
@@ -1,0 +1,18 @@
+//
+//  MainSplitViewController.swift
+//  Timeline
+//
+//  Created by Christopher Boynton on 9/20/17.
+//  Copyright © 2017 Self. All rights reserved.
+//
+
+import Cocoa
+
+class MainSplitViewController: NSSplitViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+    }
+    
+}
