+++ Timeline/View/MainWindow/MainWindowController.swift
@@ -1,0 +1,19 @@
+//
+//  MainWindowController.swift
+//  Timeline
+//
+//  Created by Christopher Boynton on 9/20/17.
+//  Copyright © 2017 Self. All rights reserved.
+//
+
+import Cocoa
+
+class MainWindowController: NSWindowController {
+
+    override func windowDidLoad() {
+        super.windowDidLoad()
+    
+        // Implement this method to handle any initialization after your window controller's window has been loaded from its nib file.
+    }
+
+}
